add styles for create, edit and delete task links

+++ resources/views/task/index.blade.php
@@ -4,30 +4,34 @@
 <div class="grid col-span-full">
     <h1 class="mb-5">@lang('views.task.index.header')</h1>
 
-    <div>
-        {{ Form::open(['route' => 'tasks.index', 'method' => 'get', 'class' => 'w-full max-w-sm']) }}
-        <div class="flex items-center mt-4">
-            <div>
-                {{ Form::select('filter[status_id]', $taskStatusesForFilterForm, Arr::get($filterQueryString, 'status_id', ''), ['class' => 'rounded border-gray-300', 'placeholder' => __('views.task.index.placeholders.status_id')]) }}
+    <div class="w-full flex items-center mt-4 ">
+        <div>
+            {{ Form::open(['route' => 'tasks.index', 'method' => 'get', 'class' => '']) }}
+            <div class="flex">
+                <div>
+                    {{ Form::select('filter[status_id]', $taskStatusesForFilterForm, Arr::get($filterQueryString, 'status_id', ''), ['class' => 'rounded border-gray-300', 'placeholder' => __('views.task.index.placeholders.status_id')]) }}
+                </div>
+                <div>
+                    {{ Form::select('filter[created_by_id]', $usersForFilterForm, Arr::get($filterQueryString, 'created_by_id', ''), ['class' => 'ml-2 rounded border-gray-300', 'placeholder' => __('views.task.index.placeholders.created_by_id')]) }}
+                </div>
+                <div>
+                    {{ Form::select('filter[assigned_to_id]', $usersForFilterForm, Arr::get($filterQueryString, 'assigned_to_id', ''), ['class' => 'ml-2 rounded border-gray-300', 'placeholder' => __('views.task.index.placeholders.assigned_to_id')]) }}
+                </div>
+                <div>
+                    {{ Form::submit(__('views.task.index.filter_button'), ['class' => 'bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded ml-2']) }}
+                </div>
+                {{ Form::close() }}
             </div>
-            <div>
-                {{ Form::select('filter[created_by_id]', $usersForFilterForm, Arr::get($filterQueryString, 'created_by_id', ''), ['class' => 'ml-2 rounded border-gray-300', 'placeholder' => __('views.task.index.placeholders.created_by_id')]) }}
-            </div>
-            <div>
-                {{ Form::select('filter[assigned_to_id]', $usersForFilterForm, Arr::get($filterQueryString, 'assigned_to_id', ''), ['class' => 'ml-2 rounded border-gray-300', 'placeholder' => __('views.task.index.placeholders.assigned_to_id')]) }}
-            </div>
-            <div>
-                {{ Form::submit(__('views.task.index.filter_button'), ['class' => 'bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded ml-2']) }}
-            </div>
-            {{ Form::close() }}
+        </div>
+
+        <div class="ml-auto">
+            @can('create', App\Models\Task::class)
+            <a href="{{ route('tasks.create') }}" class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded ml-2">
+                @lang('views.task.index.create_button')
+            </a>
+            @endcan
         </div>
     </div>
-
-    @can('create', App\Models\Task::class)
-        <a href="{{ route('tasks.create') }}">
-            @lang('views.task.index.create_button')
-        </a>
-    @endcan
 
     <table class="mt-4">
         <thead class="border-b-2 border-solid border-black text-left">
@@ -39,41 +43,38 @@
                 <th>@lang('views.task.index.assigned_to')</th>
                 <th>@lang('views.task.index.created_at')</th>
                 @can('seeActions', App\Models\Task::class)
-                    <th>@lang('views.task_status.index.actions')</th>
+                <th>@lang('views.task_status.index.actions')</th>
                 @endcan
             </tr>
         </thead>
         @foreach($tasks as $task)
-            <tr class="border-b border-dashed text-left">
-                <td>{{ $task->id }}</td>
-                <td>{{ $task->status->name }}</td>
-                <td>
-                    <a  class="text-blue-600 hover:text-blue-900" href="{{ route('tasks.show', $task) }}">
-                        {{ $task->name }}
-                    </a>
-                </td>
-                <td>{{ $task->createdBy->name }}</td>
-                <td>{{ $task->assignedTo->name ?? '' }}</td>
-                <td>{{ $task->created_at->format('d.m.Y') }}</td>
-                <td>
-                    @can('delete', $task)
-                        <a
-                            data-confirm="@lang('views.task.index.delete_confirmation')"
-                            data-method="delete"
-                            href="{{ route('tasks.destroy', $task) }}"
-                        >
-                            @lang('views.task.index.delete')
-                        </a>
-                    @endcan
-                    @can('update', $task)
-                        <a href="{{ route('tasks.edit', $task) }}">
-                            @lang('views.task.index.edit')
-                        </a>
-                    @endcan
-                </td>
-            </tr>
+        <tr class="border-b border-dashed text-left">
+            <td>{{ $task->id }}</td>
+            <td>{{ $task->status->name }}</td>
+            <td>
+                <a class="text-blue-600 hover:text-blue-900" href="{{ route('tasks.show', $task) }}">
+                    {{ $task->name }}
+                </a>
+            </td>
+            <td>{{ $task->createdBy->name }}</td>
+            <td>{{ $task->assignedTo->name ?? '' }}</td>
+            <td>{{ $task->created_at->format('d.m.Y') }}</td>
+            <td>
+                @can('delete', $task)
+                <a data-confirm="@lang('views.task.index.delete_confirmation')" data-method="delete" href="{{ route('tasks.destroy', $task) }}" class="text-red-600 hover:text-red -900">
+                    @lang('views.task.index.delete')
+                </a>
+                @endcan
+                @can('update', $task)
+                <a href="{{ route('tasks.edit', $task) }}" class="text-blue-600 hover:text-blue-900">
+                    @lang('views.task.index.edit')
+                </a>
+                @endcan
+            </td>
+        </tr>
         @endforeach
     </table>
+
     <div class="mt-4">
         {{ $tasks->links() }}
     </div>
